Fix partial (sliced) writes to receive buffers in thread backend

+++ backends/mpi/world.hpp
@@ -154,7 +154,7 @@ class world : public bulk::world {
 
     // Size is per element
     void put_(int processor, const void* values, size_t size, int var_id,
-              size_t offset, int count) override final {
+              size_t offset, size_t count) override final {
         put_buffers_[processor] << put_t::multiple;
         put_buffers_[processor] << var_id;
         size_t total_offset = offset * size;
@@ -174,7 +174,7 @@ class world : public bulk::world {
 
     // Size is per element
     void get_(int processor, int var_id, size_t size, void* target,
-              size_t offset, int count) override final {
+              size_t offset, size_t count) override final {
         get_request_buffers_[processor] << get_t::request_multiple;
         get_request_buffers_[processor] << var_id;
         size_t total_offset = offset * size;
